+++ pdata/internal/generated_wrapper_traces.go
@@ -234,7 +234,7 @@ func GenerateTestSpan() Span {
 
 func FillTestSpan(tv Span) {
 	tv.orig.TraceId = data.NewTraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1})
-	tv.orig.SpanId = data.NewSpanID([8]byte{1, 2, 3, 4, 5, 6, 7, 8})
+	tv.orig.SpanId = data.NewSpanID([8]byte{8, 7, 6, 5, 4, 3, 2, 1})
 	tv.orig.TraceState = "congo=congos"
 	tv.orig.ParentSpanId = data.NewSpanID([8]byte{8, 7, 6, 5, 4, 3, 2, 1})
 	tv.orig.Name = "test_name"
@@ -303,7 +303,7 @@ func GenerateTestSpanLink() SpanLink {
 
 func FillTestSpanLink(tv SpanLink) {
 	tv.orig.TraceId = data.NewTraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1})
-	tv.orig.SpanId = data.NewSpanID([8]byte{1, 2, 3, 4, 5, 6, 7, 8})
+	tv.orig.SpanId = data.NewSpanID([8]byte{8, 7, 6, 5, 4, 3, 2, 1})
 	tv.orig.TraceState = "congo=congos"
 	FillTestMap(NewMap(&tv.orig.Attributes))
 	tv.orig.DroppedAttributesCount = uint32(17)