+++ pdata/pcommon/spanid_test.go
@@ -20,36 +20,34 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestEmptySpanID(t *testing.T) {
-	sid := EmptySpanID
-	assert.Equal(t, [8]byte{}, sid.Bytes())
-	assert.True(t, sid.IsEmpty())
-}
-
 func TestNewSpanID(t *testing.T) {
-	sid := NewSpanID([8]byte{1, 2, 3, 4, 4, 3, 2, 1})
-	assert.Equal(t, [8]byte{1, 2, 3, 4, 4, 3, 2, 1}, sid.Bytes())
+	sid := SpanID([8]byte{1, 2, 3, 4, 4, 3, 2, 1})
+	assert.Equal(t, [8]byte{1, 2, 3, 4, 4, 3, 2, 1}, [8]byte(sid))
 	assert.False(t, sid.IsEmpty())
 }
 
+func TestNewSpanIDEmpty(t *testing.T) {
+	sid := NewSpanIDEmpty()
+	assert.Equal(t, [8]byte{}, [8]byte(sid))
+	assert.True(t, sid.IsEmpty())
+}
+
 func TestSpanIDHexString(t *testing.T) {
-	sid := NewSpanID([8]byte{})
+	sid := SpanID([8]byte{})
 	assert.Equal(t, "", sid.HexString())
 
-	sid = NewSpanID([8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23})
+	sid = SpanID([8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23})
 	assert.Equal(t, "1223ad1223ad1223", sid.HexString())
 }
 
 func TestSpanIDImmutable(t *testing.T) {
 	initialBytes := [8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23}
-	sid := NewSpanID(initialBytes)
-	assert.Equal(t, initialBytes, sid.Bytes())
+	sid := SpanID(initialBytes)
+	assert.Equal(t, SpanID(initialBytes), sid)
 
 	// Get the bytes and try to mutate.
-	bytes := sid.Bytes()
-	bytes[4] = 0x89
+	sid[4] = 0x89
 
 	// Does not change the already created SpanID.
-	assert.NotEqual(t, bytes, sid.Bytes())
-	assert.Equal(t, initialBytes, sid.Bytes())
+	assert.NotEqual(t, SpanID(initialBytes), sid)
 }